requirements.txt
translators==5.5.5
openai==0.19.0
python-dotenv==0.19.2